requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.3.0
MarkupSafe==3.0.3
numpy==2.4.0
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.4
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
scikit-learn>=1.1.0
mlflow>=2.0.0
dvc>=3.0.0
nltk>=3.8.1
matplotlib>=3.7.0
seaborn>=0.12.0
# SURGICAL ADDITIONS for ML & Dashboard
scikit-surprise>=1.1.3
plotly>=5.15.0
streamlit>=1.25.0
